requirements.txt
fastapi
uvicorn[standard]
sqlmodel
python-multipart
python-jose[cryptography]
passlib[bcrypt]
aiofiles
pydantic-settings
email-validator
jinja2
